// src/types/cart-response.ts
export interface CartResponse{
    status:string
    numOfCartItems : number
    cartId: string
    data: CartItem
}
export interface CartItem{
    _id: string
    cartOwner:string
    products: Product[]
    createdAt: string
    updatedAt: string
    __v: number
    totalCartPrice: number 
}

export interface Product{
    count: number
    _id: string
    product: Product2
    price: number
}

export interface Product2{
    subcategory : Subcategory[]
    _id: string
    title: string
    quantity: number
    imageCover: string
    brand: Brand
    ratingsAverage: number
    id: string
}

export interface Subcategory{
    _id: string
    name: string
    slug: string
    category : string
}
export interface Brand{
    _id: string
    name: string
    slug: string
    image : string
}
export type shipping = {
    details : string ,
    phone: string , 
    city : string

}